Removed unused commands
