--- server/spark/src/main/scala/com/eigengo/lift/spark/jobs/PrintSequence.scala
@@ -1,29 +1,0 @@
-package com.eigengo.lift.spark.jobs
-
-import com.typesafe.config.Config
-import org.apache.spark.{SparkConf, SparkContext}
-
-import java.net._
-
-case class PrintSequence() extends Batch[Int, Unit] {
-
-  override def name: String = "PrintSequence"
-
-  override def execute(master: String, config: Config, params: Int): Either[String, Unit] = {
-
-    println("SPARK DRIVER HOST:")
-    println(InetAddress.getLocalHost.getHostAddress)
-
-   /* val sc = new SparkContext(new SparkConf()
-      .setAppName(name)
-      .set("spark.driver.host", InetAddress.getLocalHost.getHostAddress)
-      .set("spark.driver.port", "9001")
-      .setMaster(master))
-
-    sc.parallelize(1 to params).map(_ + 1).foreach(println)
-
-    sc.stop()*/
-
-    Right()
-  }
-}
